fix faulty script exit status (SOFTWARE-4719)
oldest gotcha in the book!

you don't want the last statement in your script to be a

  [[ ... ]] && do something

because if the condition fails, then $? is set to 1, which ends up being
the script exit status.

you can fix this by changing it to

  [[ ! ... ]] || do something

although in this case i think it makes it awkward to read.

i prefer two lines to six, but in this case the long-winded if-blocks
are more readable and avoid this issue with '&&'.

the exit status of the script mainly matters because, even though the
service will complete its work regardless, systemd will be tricked into
thinking it does not work.