move scripts out of harm's way (SOFTWARE-4719)
in particular, the setup script is the only one intended to be invoked
by hand, and that by an administrator.  So put it under /usr/sbin.

Since the other two are not intended for manual inocation, tuck them
under /usr/libexec, where users won't be so easily tricked into running
them.