fix permissions handling for pwfile in sub-process
The "cat /dev/fd/9" implies re-opening the pwfile (in linux), which is
subject to the permissions check. This probably only ever worked before
if pwfile was root/644 or osg-token-svc/600.

Fix by using a dup instead:

  # ls -l testfile
  -rw------- 1 root root 7 Apr  7 17:24 testfile
  # cat testfile
  winner
  # exec 9< testfile
  # su - user
  $ cat /dev/fd/9
  cat: /dev/fd/9: No such file or directory
  $ cat < /dev/fd/9
  -su: /dev/fd/9: Permission denied
  $ cat <&9
  winner

In order to handle this sensibly in the script, add a --pw-fd option,
for internal use only, as --pw-file /dev/fd/9 is fundamentally broken
since it can't be opened by the (non-root) service account.