use oidc-add --pw-file option from newer versions (SOFTWARE-5050)
Note this works, like the --pw-cmd option used to, because this script
is invoked by systemd, and the service file sets up Capabilities for
the osg-token-svc account to be able to skip permissions checks...

Unfortunately the same is not true for the setup script, which is run
by hand as root, and only is able to provide access to the password
file via an inherited file descriptor.  Thus re-opening the file path
itself (which is implied by --pw-file) will not work when called in
the setup script after it is re-invoked as the service account.