rejigger pw file checks (SOFTWARE-5304)
the [[ -e $pwfile ]] check should only be done when running as root,
which opens $pwfile, and then su-exec's into the non-root service
account to awake with the pw fd open.  The non-root service account
does not have the need, nor necessarily the ability, to check whether
the pwfile actually exists -- the check will fail if the tokens dir
has 0700 perms, as the service account won't be able to do a directory
search in that case.  The only thing the service account needs is to
read from its open fd.

Updating the checks accordingly.