move the pwfile redirection _inside_ the 'if' block (SOFTWARE-5304)
ehhhh... the redirection was outside the closing 'fi' so that it
would apply to the whole 'if' block.  Which _includes_ the check
in the 'if' condition.  But in the service account case (when $UID != 0)
pwfile is likely not readable, which means the redirection for the whole
'if' block will fail _before_ the [[ $UID = 0 ]] check even gets a
chance to see the light of day.  (Or the stars of night, whatever.)

So really the [[ $UID = 0 ]] check needs to run _before_ the pwfile
redirection; thus we move it to inside the 'if' block.